Extract singleton utils to its own module
