+++ src/orion/client/__init__.py
@@ -12,8 +12,8 @@ from orion.client.cli import (
     interrupt_trial, report_bad_trial, report_objective, report_results)
 from orion.client.experiment import ExperimentClient
 import orion.core.io.experiment_builder as experiment_builder
-from orion.core.utils import update_singletons
 from orion.core.utils.exceptions import RaceCondition
+from orion.core.utils.singleton import update_singletons
 from orion.core.worker.producer import Producer
 from orion.storage.base import setup_storage
